Split the sentence into words(strtok)
Write a program that asks the user to enter a sentence and separates this sentence into words and saves the words into a string array. The sentence can be limited to 20 words. Then save the words in a text file, one after the other. The user may or may not have placed a period at the end of the sentence. The code you write must provide both conditions.